+++ tests/unit/background.spec.js
@@ -1,17 +1,8 @@
 require("./util");
 
-const _ = require("lodash");
-import LocalStore from "../../src/js/shared/local_store";
-import { StoreKey } from "../../src/js/shared/constants";
-import { getProviderTargetURL } from "../../src/js/shared/misc";
-
 const ConfigFile = require("../../src/js/shared/config_file");
 const { MiscURLs } = require("../../src/js/shared/constants");
-const {
-  ContextualMenu,
-  installedListener,
-  onClickedListener,
-} = require("../../src/background");
+const { installedListener } = require("../../src/background");
 
 const createTabs = jest.spyOn(chrome.tabs, "create");
 const sanitizeSettings = jest.spyOn(ConfigFile.default, "sanitizeSettings");
@@ -41,205 +32,5 @@ describe("background.js", () => {
     await installedListener({ reason: "install" });
     expect(sanitizeSettings).toHaveReturned();
     expect(updateNow).toHaveBeenCalled();
-  });
-});
-
-describe("ContextMenu", () => {
-  beforeAll(async () => {
-    let rsa = await LocalStore.getOne(StoreKey.RSA_SECURITY);
-    let nwi = await LocalStore.getOne(StoreKey.NET_WITNESS);
-    let cbc = await LocalStore.getOne(StoreKey.CARBON_BLACK);
-    rsa.config.RSAConfigEnable = true;
-    nwi.config.NWIConfigEnable = true;
-    cbc.config.CBCConfigEnable = true;
-    await LocalStore.setOne(StoreKey.RSA_SECURITY, rsa);
-    await LocalStore.setOne(StoreKey.NET_WITNESS, nwi);
-    await LocalStore.setOne(StoreKey.CARBON_BLACK, cbc);
-
-    // Update contextual menu.
-    await ContextualMenu.update();
-  });
-
-  it("Should open tab for simple search provider", async () => {
-    const info = {
-      menuItemId: "searchprovider-0",
-      selectionText: "test",
-    };
-    const tab = {
-      index: 10,
-    };
-    await onClickedListener(info, tab);
-
-    const providers = await LocalStore.getOne(StoreKey.SEARCH_PROVIDERS);
-    const settings = await LocalStore.getOne(StoreKey.SETTINGS);
-    const provider = _.find(providers, function (item) {
-      return item.menuIndex === info.menuItemId;
-    });
-    const targetURL = getProviderTargetURL(provider, info.selectionText);
-    expect(createTabs).toBeCalledWith({
-      url: targetURL,
-      selected: !settings.resultsInBackgroundTab,
-      index: tab.index + 1,
-    });
-  });
-
-  it("Should open dialog for POST search provider", async () => {
-    const info = {
-      menuItemId: "searchprovider-0",
-      selectionText: "test",
-    };
-    const tab = {
-      index: 10,
-    };
-    const providers = await LocalStore.getOne(StoreKey.SEARCH_PROVIDERS);
-    const settings = await LocalStore.getOne(StoreKey.SETTINGS);
-    providers[0].postEnabled = true;
-    await LocalStore.setOne(StoreKey.SEARCH_PROVIDERS, providers);
-    await onClickedListener(info, tab);
-    const targetURL = getProviderTargetURL(providers[0], info.selectionText);
-
-    expect(createTabs).toBeCalledWith({
-      url: targetURL,
-      selected: !settings.resultsInBackgroundTab,
-      index: tab.index + 1,
-    });
-  });
-
-  it("Should open dialog for proxy search provider", async () => {
-    const info = {
-      menuItemId: "searchprovider-0",
-      selectionText: "test",
-    };
-    const tab = {
-      index: 10,
-    };
-    const providers = await LocalStore.getOne(StoreKey.SEARCH_PROVIDERS);
-    const settings = await LocalStore.getOne(StoreKey.SETTINGS);
-    providers[0].proxyEnabled = true;
-    providers[0].proxyUrl =
-      "https://run.mocky.io/v3/49340c84-b278-4332-8c5c-59ef778fb958";
-    await LocalStore.setOne(StoreKey.SEARCH_PROVIDERS, providers);
-    await onClickedListener(info, tab);
-    const targetURL = getProviderTargetURL(providers[0], info.selectionText);
-
-    expect(createTabs).toBeCalledWith({
-      url: targetURL,
-      selected: !settings.resultsInBackgroundTab,
-      index: tab.index + 1,
-    });
-  });
-
-  it("Should open group providers", async () => {
-    const newProviders = [
-      {
-        menuIndex: "searchprovider-37",
-        label: "Provider belonging to group 3",
-        link: "https://stat.ripe.net/TESTSEARCH#tabId=at-a-glance",
-        enabled: true,
-        fromConfig: true,
-        group: 8,
-        postEnabled: false,
-        postValue: "",
-        proxyEnabled: false,
-        proxyUrl: "",
-      },
-      {
-        menuIndex: "searchprovider-38",
-        label: "Provider 2 belonging to group 3",
-        link: "http://www.google.com/safebrowsing/diagnostic?site=TESTSEARCH",
-        enabled: true,
-        fromConfig: true,
-        group: 8,
-        postEnabled: false,
-        postValue: "",
-        proxyEnabled: false,
-        proxyUrl: "",
-      },
-    ];
-    let providers = await LocalStore.getOne(StoreKey.SEARCH_PROVIDERS);
-    providers = providers.concat(newProviders);
-    await LocalStore.setOne(StoreKey.SEARCH_PROVIDERS, providers);
-
-    const settings = await LocalStore.getOne(StoreKey.SETTINGS);
-    const info = {
-      menuItemId: "group-3",
-      selectionText: "test",
-    };
-    const tab = {
-      index: 10,
-    };
-    const urls = _.map(newProviders, function (provider) {
-      return getProviderTargetURL(provider, info.selectionText);
-    });
-
-    // If should open in new window.
-    await onClickedListener(info, tab);
-    expect(
-      chrome.windows.create.withArgs({
-        url: urls,
-        focused: !settings.resultsInBackgroundTab,
-      }).calledOnce
-    ).toBe(true);
-
-    // If should open in new tabs.
-    settings.openGroupsInNewWindow = false;
-    await LocalStore.setOne(StoreKey.SETTINGS, settings);
-    await onClickedListener(info, tab);
-    expect(createTabs).toBeCalledWith({
-      url: urls[0],
-      selected: !settings.resultsInBackgroundTab,
-      index: settings.enableAdjacentTabs ? ++tab.index : null,
-    });
-    expect(createTabs).toBeCalledWith({
-      url: urls[1],
-      selected: !settings.resultsInBackgroundTab,
-      index: settings.enableAdjacentTabs ? ++tab.index : null,
-    });
-  });
-
-  it("Should open options page", async () => {
-    await onClickedListener({ menuItemId: "optionspage" });
-    expect(chrome.runtime.openOptionsPage.calledOnce).toBe(true);
-  });
-
-  it("Should open RSA item", async () => {
-    const info = {
-      parentMenuItemId: "rsasecurity-0",
-      menuItemId: "rsasecurity-0_1",
-      selectionText: "test",
-    };
-    await onClickedListener(info);
-    var subUrl = "https://192.168.1.10/investigation/2/navigate/query";
-    expect(createTabs).toBeCalledWith({
-      selected: true,
-      url: expect.stringContaining(subUrl),
-    });
-  });
-
-  it("Should open CBC item", async () => {
-    const info = {
-      parentMenuItemId: "parent-1",
-      menuItemId: "carbonblack-0",
-      selectionText: "test",
-    };
-    await onClickedListener(info);
-    var subUrl = "https://192.168.1.10/#/search/cb.urlver=1&q=test";
-    expect(createTabs).toBeCalledWith({
-      selected: true,
-      url: expect.stringContaining(subUrl),
-    });
-  });
-
-  it("Should open NWI item", async () => {
-    const info = {
-      parentMenuItemId: "netwitness-0",
-      menuItemId: "netwitness-0_1",
-      selectionText: "test",
-    };
-    await onClickedListener(info);
-    var subUrl = "nw:///?collection=&where=%28alias.host%3D'test";
-    expect(createTabs).toBeCalledWith({
-      url: expect.stringContaining(subUrl),
-    });
   });
 });
